The Kernel.RunAsync instance method is moved to extension method.
