+++ dotnet/src/SemanticKernel.Core/KernelExtensions.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
@@ -166,5 +167,128 @@ public static class KernelExtensions
     {
         Verify.NotNull(kernel);
         return kernel.RunAsync(new ContextVariables(input), cancellationToken, pipeline);
+    }
+
+    /// <summary>
+    /// Run a pipeline composed of synchronous and asynchronous functions.
+    /// </summary>
+    /// <param name="kernel">The kernel.</param>
+    /// <param name="variables">Input to process</param>
+    /// <param name="cancellationToken">The <see cref="CancellationToken"/> to monitor for cancellation requests. The default is <see cref="CancellationToken.None"/>.</param>
+    /// <param name="pipeline">List of functions</param>
+    /// <returns>Result of the function composition</returns>
+    /// <inheritdoc/>
+    public static async Task<KernelResult> RunAsync(this Kernel kernel, ContextVariables variables, CancellationToken cancellationToken, params ISKFunction[] pipeline)
+    {
+        var context = kernel.CreateNewContext(variables);
+
+        FunctionResult? functionResult = null;
+
+        int pipelineStepCount = 0;
+        var allFunctionResults = new List<FunctionResult>();
+
+        var logger = kernel.LoggerFactory.CreateLogger(typeof(Kernel));
+
+        foreach (ISKFunction skFunction in pipeline)
+        {
+repeat:
+            cancellationToken.ThrowIfCancellationRequested();
+
+            try
+            {
+                var functionDetails = skFunction.Describe();
+
+                functionResult = await skFunction.InvokeAsync(context, null, cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                if (IsCancelRequested(skFunction, functionResult.Context, pipelineStepCount, logger))
+                {
+                    break;
+                }
+
+                if (IsSkipRequested(skFunction, functionResult.Context, pipelineStepCount, logger))
+                {
+                    continue;
+                }
+
+                // Only non-stop results are considered as Kernel results
+                allFunctionResults.Add(functionResult!);
+
+                if (IsRepeatRequested(skFunction, functionResult.Context, pipelineStepCount, logger))
+                {
+                    goto repeat;
+                }
+            }
+            catch (Exception ex)
+            {
+                logger.LogError("Plugin {Plugin} function {Function} call fail during pipeline step {Step} with error {Error}:", skFunction.PluginName, skFunction.Name, pipelineStepCount, ex.Message);
+                throw;
+            }
+
+            pipelineStepCount++;
+        }
+
+        return KernelResult.FromFunctionResults(allFunctionResults.LastOrDefault()?.Value, allFunctionResults);
+    }
+
+    /// <summary>
+    /// Checks if the handler requested to cancel the function execution.
+    /// </summary>
+    /// <param name="skFunction">Target function</param>
+    /// <param name="context">Context of execution</param>
+    /// <param name="pipelineStepCount">Current pipeline step</param>
+    /// <param name="logger">The logger.</param>
+    /// <returns></returns>
+    private static bool IsCancelRequested(ISKFunction skFunction, SKContext context, int pipelineStepCount, ILogger logger)
+    {
+        if (SKFunction.IsInvokingCancelRequested(context))
+        {
+            logger.LogInformation("Execution was cancelled on function invoking event of pipeline step {StepCount}: {PluginName}.{FunctionName}.", pipelineStepCount, skFunction.PluginName, skFunction.Name);
+            return true;
+        }
+
+        if (SKFunction.IsInvokedCancelRequested(context))
+        {
+            logger.LogInformation("Execution was cancelled on function invoked event of pipeline step {StepCount}: {PluginName}.{FunctionName}.", pipelineStepCount, skFunction.PluginName, skFunction.Name);
+            return true;
+        }
+
+        return false;
+    }
+
+    /// <summary>
+    /// Checks if the handler requested to skip the function execution.
+    /// </summary>
+    /// <param name="skFunction">Target function</param>
+    /// <param name="context">Context of execution</param>
+    /// <param name="pipelineStepCount">Current pipeline step</param>
+    /// <param name="logger">The logger.</param>
+    /// <returns></returns>
+    private static bool IsSkipRequested(ISKFunction skFunction, SKContext context, int pipelineStepCount, ILogger logger)
+    {
+        if (SKFunction.IsInvokingSkipRequested(context))
+        {
+            logger.LogInformation("Execution was skipped on function invoking event of pipeline step {StepCount}: {PluginName}.{FunctionName}.", pipelineStepCount, skFunction.PluginName, skFunction.Name);
+            return true;
+        }
+
+        return false;
+    }
+
+    /// <summary>
+    /// Checks if the handler requested to repeat the function execution.
+    /// </summary>
+    /// <param name="skFunction">Target function</param>
+    /// <param name="context">Context of execution</param>
+    /// <param name="pipelineStepCount">Current pipeline step</param>
+    /// <param name="logger">The logger.</param>
+    /// <returns></returns>
+    private static bool IsRepeatRequested(ISKFunction skFunction, SKContext context, int pipelineStepCount, ILogger logger)
+    {
+        if (context.FunctionInvokedHandler?.EventArgs?.IsRepeatRequested ?? false)
+        {
+            logger.LogInformation("Execution repeat request on function invoked event of pipeline step {StepCount}: {PluginName}.{FunctionName}.", pipelineStepCount, skFunction.PluginName, skFunction.Name);
+            return true;
+        }
+        return false;
     }
 }
